this is how I feel now that im done
Then they (we) cried out to the Lord in their trouble, and he brought them out of their distress. He stilled the storm to a whisper(its ending); the waves of the sea were hushed(free weekend). They were glad when it grew calm, and he guided them to their desired haven.” (Psalm 107:28-30)